Use for loop instead of for_each
I believe for_each was making this more complicated than it needed to
be, and an ordinary for loop saves a level of indentation anyway. The
for_each method is typically for when your iterator chain is big and
the loop body is small which is the reverse of what we have here.